+++ include/dp_flow.h
@@ -127,7 +127,7 @@ int dp_build_flow_key(struct flow_key *key /* out */, struct rte_mbuf *m /* in *
 void dp_invert_flow_key(const struct flow_key *key /* in */, struct flow_key *inv_key /* out */);
 int dp_flow_init(int socket_id);
 void dp_flow_free(void);
-void dp_process_aged_flows(int port_id);
+void dp_process_aged_flows(uint16_t port_id);
 void dp_process_aged_flows_non_offload(void);
 void dp_free_flow(struct dp_ref *ref);
 void dp_free_network_nat_port(const struct flow_value *cntrack);
